requirements.txt
feedparser
beautifulsoup4
sqlalchemy
requests
python-dotenv
fastapi
celery
uvicorn
psycopg2-binary